+++ resources/src/stores/index.js
@@ -2,13 +2,13 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 import createPersistedState from 'vuex-persistedstate';
 
-import login from '@/stores/modules/Login';
+import auth from '@/stores/modules/Auth';
 
 Vue.use(Vuex);
 
 export default new Vuex.Store({
   modules: {
-    login,
+    auth,
   },
   plugins: [createPersistedState({ storage: window.sessionStorage })],
 })
